backend/requirements.txt
# Core - Using stable versions from mid-2023
fastapi==0.100.0
uvicorn[standard]==0.23.1
python-dotenv==1.0.0
pydantic==2.1.1
pydantic-settings==2.0.2
email-validator==2.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
azure-cosmos==4.5.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-cov==4.1.0
pytest-mock==3.11.1
faker==19.2.0

# Development
black==23.7.0
flake8==6.1.0
mypy==1.4.1
pre-commit==3.3.3

# AI/Voice (Phase 3)
openai==1.8.0  # For Whisper voice transcription
anthropic==0.8.1  # For Claude text structuring